Widen IUniswapV3Pool.slot0's feeProtocol to uint32, unblocking the launch
Confirmed against BSC mainnet: the PancakeSwap V3 WBNB/USDT 0.25% pool
(0x1401ff943D08a7E098328C1d3a9d388923B115D2) reports feeProtocol 209718400,
which is 3200 | (3200 << 16) — a 32% protocol fee on both sides. The interface
declared uint8, so the ABI decoder rejected that word and _getInitialSqrtPriceX96
reverted, taking the whole launch with it on its first call.

uint32 is PancakeSwap's actual width and decodes Uniswap pools too, since their
uint8 value fits. The field is read and discarded either way — the price math
never uses it — and the tests assert the launch price is byte-identical across
both pool shapes, so this is a decoding fix with no behavioural change.

The compatibility tests now pin the working direction rather than the broken one,
with the live mainnet value as a regression case.

+++ contracts/UniswapV3TokenInitializer.sol
@@ -51,7 +51,12 @@ interface IUniswapV3Factory {
 
 interface IUniswapV3Pool {
     function tickSpacing() external view returns (int24);
-    function slot0() external view returns (uint160 sqrtPriceX96, int24 tick, uint16 observationIndex, uint16 observationCardinality, uint16 observationCardinalityNext, uint8 feeProtocol, bool unlocked);
+    // feeProtocol is uint8 on Uniswap V3 but uint32 on PancakeSwap V3, which packs it as
+    // feeProtocol0 | (feeProtocol1 << 16). Solidity's ABI decoder rejects a word that does
+    // not fit the declared type, so uint8 here reverts against any Pancake pool carrying a
+    // protocol fee — BSC's WBNB/USDT 0.25% pool reads 209718400 (3200 on both sides). The
+    // wider type decodes both, since a Uniswap pool's uint8 value fits in a uint32.
+    function slot0() external view returns (uint160 sqrtPriceX96, int24 tick, uint16 observationIndex, uint16 observationCardinality, uint16 observationCardinalityNext, uint32 feeProtocol, bool unlocked);
 }
 
 contract UniswapV3TokenInitializer {
